tile, serialize/save: limit tile atlas size to 0xff

+++ src/tile-image.hpp
@@ -6,7 +6,7 @@ namespace floormat {
 
 struct tile_atlas;
 
-using variant_t = std::uint16_t;
+using variant_t = std::uint8_t;
 
 struct tile_image_proto final
 {
